% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_add_main_url.R
\name{hd_add_main_url}
\alias{hd_add_main_url}
\title{Paste the main page url for homedepot.com to a character string of a partial url}
\usage{
hd_add_main_url(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Scraping all product links from catalog pages returns partial links, this function appends the main url as a prefix.
}
\description{
Paste the main page url for homedepot.com to a character string of a partial url
}
\examples{
\dontrun{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links_part <- hd_get_product_links(x)
links_full <- hd_add_main_url(links)
}
}

