% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_build_multi_page_links.R
\name{hd_build_multi_page_links}
\alias{hd_build_multi_page_links}
\title{Construct the urls for the multiple return pages for a given catalog}
\usage{
hd_build_multi_page_links(x, pages = max_page)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}

\item{pages}{An integer, the number of pages included in the catalog (default is to find automatically but you can override)}
}
\value{
Returns of vector of character strings containing urls for all pages of search results for a catalog of products on HD
}
\description{
Construct the urls for the multiple return pages for a given catalog
}
\examples{
\dontrun
{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_build_multi_page_links(x)
all_prods_in_cat <- map(links, hd_get_product_links)
}
}

