% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_catalog_scrape.R
\name{hd_catalog_scrape}
\alias{hd_catalog_scrape}
\title{Get product and price details for all products in a section on Home Depot from main url}
\usage{
hd_catalog_scrape(x)
}
\arguments{
\item{links}{A list of character strings for web urls}
}
\value{
Returns a single data frame with data for every product within a catalog section
}
\description{
Get product and price details for all products in a section on Home Depot from main url
}
\examples{
\dontrun{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
hd_catalog_scrape(x)
}
}

