% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_find_category_from_sitemap.R
\name{hd_find_category_from_sitemap}
\alias{hd_find_category_from_sitemap}
\title{Use a string to search product categories on home depot's site map}
\usage{
hd_find_category_from_sitemap(search_string)
}
\arguments{
\item{search_string}{A character string to use to find relevant urls}
}
\value{
Returns a list of urls of product categories containing your search string
}
\description{
Use a string to search product categories on home depot's site map
}
\examples{
\dontrun{
page_links <- hd_find_category_from_sitemap("Electrical-Wire")[-1]
prod_links <- map(page_links, hd_build_multi_page_links) \%>\%
  setNames(., map(., ~slit(.[[1]], pos = 5)))
}
}

