% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_get_product_links.R
\name{hd_get_product_links}
\alias{hd_get_product_links}
\title{Scrape all product urls from a catalog page on HD}
\usage{
hd_get_product_links(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Returns a vector of strings with product urls
}
\description{
Scrape all product urls from a catalog page on HD
}
\examples{
\dontrun{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_get_product_links(x) \%>\%
hd_add_main_url \%>\%
map(hd_tidy_price) \%>\%
reduce(rbind)
}
}

