% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_product_model.R
\name{hd_product_model}
\alias{hd_product_model}
\title{Scrape and format model number for a given product from HD product page}
\usage{
hd_product_model(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Returns the model number from a given HD product link
}
\description{
Scrape and format model number for a given product from HD product page
}
\examples{
\dontrun{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_get_product_links(x)
hd_product_model(links[1])
}
}

