% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_product_specs.R
\name{hd_product_specs}
\alias{hd_product_specs}
\title{Scrape and format product specifications from a product's web page}
\usage{
hd_product_specs(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Returns a list of tibbles of product characteristics and specs
}
\description{
Scrape and format product specifications from a product's web page
}
\examples{
\dontrun{
rx142 <- "http://www.homedepot.com/p/Cerrowire-250-ft-14-2-White-NM-B-Wire-147-1472G/202304609"
hd_product_specs(rx142)
}
}

