% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_tidy_price.R
\name{hd_tidy_price}
\alias{hd_tidy_price}
\title{Get product and price details from HD web page in a tidy format}
\usage{
hd_tidy_price(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Returns a data frame with the product's model number and price from web
}
\description{
Get product and price details from HD web page in a tidy format
}
\examples{
\dontrun{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_get_product_links(x)
map(links, hd_tidy_price) \%>\% reduce(rbind)
}
}

