% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_tidy_scrape.R
\name{hd_tidy_scrape}
\alias{hd_tidy_scrape}
\title{Get product and price details for a list of product urls on Home Depot}
\usage{
hd_tidy_scrape(links)
}
\arguments{
\item{links}{A list of character strings for web urls}
}
\value{
Returns a list of data frames with the product's model number and price from web
}
\description{
Get product and price details for a list of product urls on Home Depot
}
\examples{
\dontrun
{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_get_product_links(x)
hd_tidy_scrape(links)
}
}

