% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_helper_for_scraping_specs.R
\name{help}
\alias{help}
\alias{is_col_head}
\alias{is_in_details_table}
\alias{is_in_dimensions_table}
\alias{is_in_product_features_table}
\alias{is_in_warranty_table}
\title{Helper functions}
\usage{
is_col_head(x)

is_in_dimensions_table(x)

is_in_details_table(x)

is_in_warranty_table(x)

is_in_product_features_table(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Helper functions
}

