% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_page.R
\name{next_page}
\alias{next_page}
\title{Extract the max page number from bottom of product catalog url}
\usage{
next_page(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Returns a numeric value containing the last page in catalog
}
\description{
Extract the max page number from bottom of product catalog url
}
\examples{
\dontrun{
pg1 <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
pg2 <- next_page(pg1)
}
}

