% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_loop.R
\name{scrape_loop}
\alias{scrape_loop}
\title{Loop a given function through a list of links and store data in a list}
\usage{
scrape_loop(links, fun)
}
\arguments{
\item{links}{A list of character strings for web urls}
}
\value{
Returns a list of results
}
\description{
Loop a given function through a list of links and store data in a list
}
\examples{
\dontrun{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_build_multi_page_links(x)
scrape_loop(links, hd_get_product_links)
}
}

