% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_product_specs.R
\name{hd_product_specs}
\alias{hd_product_specs}
\title{Scrape and format product specifications from a product link on Home Depot.com}
\usage{
hd_product_specs(x)
}
\arguments{
\item{x}{A string, a connection, or a raw vector.}
}
\value{
Returns a data frame of product characteristics and specs
}
\description{
Scrape and format product specifications from a product link on Home Depot.com
}
\examples{
\dontrun
{
x <- "http://www.homedepot.com/b/Electrical-Dimmers-Switches-Outlets/N-5yc1vZc34h"
links <- hd_get_product_links(x)
hd_product_specs(links[1])
}
}

